Make the SequenceInput component work
Attempt made to build a list of {key, reducer} records
that can be passed to the stateless SequenceInput
component from the parent Simon component. Allows
that to take advantage of lifecycle events when needed,
and removed more rendering code from the main Simon
component, which should only be concerned with advancing
game state, and defining the trigger actions for the
transitions between.

The way record passed seems wrong: a type of `inputModel`
is defined in the Simon module, which describes the record.
This is then passed to the child as `Simon.inputModel`. That works,
but seems to create a circular dependency - ninja is screaming about it,
but it all builds.

Second issue is this: the callback has a type of `eactEventRe.Mouse.t =>
unit`. This, well, this is not correct. Again, still works.